FilesTransferrer_one_direction/requirements.txt
asgiref==3.4.1
click==8.0.1
fastapi==0.66.0
h11==0.12.0
pydantic==1.8.2
python-multipart==0.0.5
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
pyqrcode==1.2.1
python-dotenv==0.20.0
pypng==0.0.21